# docs/swagger.yaml
basePath: /api
definitions:
...
  entity.TModLoaderCluster:
    properties:
      auto_create:
        enum:
        - 1
        - 2
        - 3
        type: integer
      created_at:
        type: string
      difficulty:
        enum:
        - 0
        - 1
        - 2
        - 3
        type: integer
      id:
        type: string
      max_players:
        type: integer
      mods:
        items:
          $ref: '#/definitions/entity.TModLoaderModInCluster'
        type: array
      password:
        type: string
      port:
        maximum: 65535
        minimum: 1
        type: integer
      seed:
        type: string
      updated_at:
        type: string
      world_name:
        type: string
    type: object
  entity.TModLoaderMod:
    properties:
      config:
        type: string
      created_at:
        type: string
      id:
        type: string
      mod_id:
        type: string
      name:
        type: string
      remark:
        type: string
      updated_at:
        type: string
    type: object
  entity.TModLoaderModInCluster:
    properties:
      created_at:
        type: string
      current_config:
        type: string
      id:
        type: string
      mod:
        $ref: '#/definitions/entity.TModLoaderMod'
      mod_id:
        type: string
      t_mod_loader_cluster_id:
        type: string
      updated_at:
        type: string
    type: object
...
  http.response:
    type: object
host: 0.0.0.0:5800
info:
  contact: {}
  description: Aurora Admin API
  title: Aurora Admin
  version: "1.0"
paths:
...
  /api/tmodloader/archive/download:
    get:
      consumes:
      - application/json
      parameters:
      - description: name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 下载存档
      tags:
      - tmodloader_cluster
  /api/tmodloader/cluster/:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.TModLoaderCluster'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 创建存档
      tags:
      - tmodloader_cluster
  /api/tmodloader/cluster/{id}:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 删除存档
      tags:
      - tmodloader_cluster
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.TModLoaderCluster'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 获取存档信息
      tags:
      - tmodloader_cluster
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.TModLoaderCluster'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.TModLoaderCluster'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 更新存档
      tags:
      - tmodloader_cluster
  /api/tmodloader/clusters:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.TModLoaderCluster'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 获取存档列表
      tags:
      - tmodloader_cluster
...
  /api/tmodloader/control/{action}:
    post:
      consumes:
      - application/json
      parameters:
      - description: '[ start | stop | restart | update | install ]'
        in: path
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 服务器控制
      tags:
      - tmodloader_control
  /api/tmodloader/mod:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 删除 Mods
      tags:
      - tmodloader_mod
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.TModLoaderMod'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 查看 Mods
      tags:
      - tmodloader_mod
    put:
      consumes:
      - application/json
      parameters:
      - description: '[ true | false ]'
        in: query
        name: update_mod_info
        type: string
      - description: body
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/entity.TModLoaderMod'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 更新 Mods
      tags:
      - tmodloader_mod
  /api/tmodloader/mod/mod_id:
    post:
      consumes:
      - text/plain
      parameters:
      - description: 多个 Mod ID 一行一个
        in: body
        name: request
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 通过 Mod ID 添加 Mods
      tags:
      - tmodloader_mod
...
swagger: "2.0"